+++ src/BotManager.php
@@ -123,12 +123,18 @@ class BotManager
         // Make sure this is a valid call.
         $this->validateSecret();
 
+        if (!$this->isValidRequest()) {
+            throw new InvalidAccessException('Invalid access');
+        }
+
         if ($this->action->isAction(['set', 'unset', 'reset'])) {
             $this->validateAndSetWebhook();
         } elseif ($this->action->isAction('handle')) {
-            // Set any extras.
             $this->setBotExtras();
             $this->handleRequest();
+        } elseif ($this->action->isAction('cron')) {
+            $this->setBotExtras();
+            $this->handleCron();
         }
 
         return $this;
@@ -338,6 +344,25 @@ class BotManager
     }
 
     /**
+     * Handle cron.
+     *
+     * @return \NPM\TelegramBotManager\BotManager
+     * @throws \Longman\TelegramBot\Exception\TelegramException
+     */
+    public function handleCron(): self
+    {
+        $groups = explode(',', $this->params->getScriptParam('g', 'default'));
+
+        $commands = [];
+        foreach ($groups as $group) {
+            $commands = array_merge($commands, $this->params->getBotParam('cron.groups.' . $group, []));
+        }
+        $this->telegram->runCommands($commands);
+
+        return $this;
+    }
+
+    /**
      * Get the number of seconds the script should loop.
      *
      * @return int
@@ -456,10 +481,6 @@ class BotManager
      */
     public function handleWebhook(): self
     {
-        if (!$this->isValidRequest()) {
-            throw new InvalidAccessException('Invalid access');
-        }
-
         $this->telegram->handle();
 
         return $this;
@@ -487,6 +508,7 @@ class BotManager
      */
     public function isValidRequest(): bool
     {
+        // If we're running from CLI, requests are always valid, unless we're running the tests.
         if ((!self::inTest() && 'cli' === PHP_SAPI) || false === $this->params->getBotParam('validate_request')) {
             return true;
         }